requirements.txt
openai>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.2
pyyaml>=6.0.1
numpy>=1.24.0
pytest>=7.4.0
typing-extensions>=4.7.1
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0
datasets>=2.14.0
huggingface_hub>=0.19.0
peft>=0.6.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.0
evaluate>=0.4.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
wandb>=0.15.0